backend/requirements.txt
fastapi
uvicorn
openai-whisper
fuzzywuzzy
python-Levenshtein
google-generativeai
